Tidying up, pushing some changes made in the tute
Reworked the debug flag so that it's all inside the Game class as a
static variable, now the other classes just refer to that.

Added a default case to the switch (whoops)

Need to stop dicking around with details and get this thing up and
running. THEN I can play with it.